Add skeleton support for Video media type VSCO has another media type "video", which is not the same as "image" media with is_video set to True.
Add proper support for multiple types of Media by creating an Interface
with methods we care about (saving, getting correct URL, upload date)
and warn if we get a media type we haven't seen before.

The video type uses a m3u8 playlist which will
require a bit more support to support saving, so keep this as a TODO for
now.

Update page size to 30 to match API with comment explaining why